Fix a couple of bugs, add new information to the output TTree produced by the PhaseITreeMaker tool

+++ DataModel/HeftyInfo.h
@@ -19,7 +19,7 @@ class HeftyInfo : public SerialisableObject {
 
     inline HeftyInfo(int sequence_id,
       const std::vector<unsigned long long>& time,
-      const std::vector<int>& label, const std::vector<int>& t_since_beam,
+      const std::vector<int>& label, const std::vector<long long>& t_since_beam,
       const std::vector<int>& more) : sequence_id_(sequence_id), time_(time),
       label_(label), t_since_beam_(t_since_beam), more_( more.back() == 1 )
     {
@@ -31,7 +31,8 @@ class HeftyInfo : public SerialisableObject {
     inline int sequence_id() const { return sequence_id_; }
     inline unsigned long long time(size_t mb) const { return time_.at(mb); }
     inline int label(size_t mb) const { return label_.at(mb); }
-    inline int t_since_beam(size_t mb) const { return t_since_beam_.at(mb); }
+    inline long long t_since_beam(size_t mb) const
+      { return t_since_beam_.at(mb); }
     inline bool more() const { return more_; }
 
     inline virtual bool Print() override {
@@ -50,7 +51,7 @@ class HeftyInfo : public SerialisableObject {
 
     // Time since the last beam (or source) trigger for the start of each
     // minibuffer
-    std::vector<int> t_since_beam_;
+    std::vector<long long> t_since_beam_;
 
     bool more_; // true if last element of More[] == 1, false otherwise
